Refactor 1: RoomForm, VnumPicker
